# source & MISC internal ref locations rebuilt
A systemic glitch in OGA v0.2.0 data elliptical tokens (xlink:href="#e_\d+") resulted in internal references in the tok01_cts01.xml files that were non-sequential or non-ascending. These needed to be corrected by having "#e" tokens inherit the internal reference locations found in the immediately prior word token. That and faulty logic in a previous script of mine led to errors constructing the internal reference locations in the MISC fields as the third subfield. Using a custom author and document abbreviation dictionary, as well as the correct internal location references of the first and last token in each sentence, I rebuilt the "# source =" rows.